refactor: move Mechanics and Components to new package
