+++ src/main/java/zone/vao/nexoAddon/NexoAddon.java
@@ -21,8 +21,8 @@ import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.scheduler.BukkitTask;
 import org.jetbrains.annotations.NotNull;
-import zone.vao.nexoAddon.components.Components;
-import zone.vao.nexoAddon.mechanics.Mechanics;
+import zone.vao.nexoAddon.items.Components;
+import zone.vao.nexoAddon.items.Mechanics;
 import zone.vao.nexoAddon.commands.NexoAddonCommand;
 import zone.vao.nexoAddon.events.PlayerCommandPreprocessListener;
 import zone.vao.nexoAddon.events.PrepareRecipesListener;
