+++ src/main/java/zone/vao/nexoAddon/items/components/SkullValue.java
@@ -1,0 +1,4 @@
+package zone.vao.nexoAddon.items.components;
+
+public record SkullValue(String value) {
+}
