+++ src/main/java/zone/vao/nexoAddon/items/mechanics/Aura.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.mechanics;
+package zone.vao.nexoAddon.items.mechanics;
 
 import org.bukkit.Particle;
