+++ src/main/java/zone/vao/nexoAddon/items/mechanics/BedrockBreak.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.mechanics;
+package zone.vao.nexoAddon.items.mechanics;
 
 public record BedrockBreak(int hardness, double probability, int durabilityCost, boolean disableOnFirstLayer) {
 }
