+++ src/main/java/zone/vao/nexoAddon/items/mechanics/BottledExp.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.mechanics;
+package zone.vao.nexoAddon.items.mechanics;
 
 import com.nexomc.nexo.api.NexoItems;
 import net.kyori.adventure.text.minimessage.MiniMessage;
