+++ src/main/java/zone/vao/nexoAddon/utils/handlers/ParticleEffectManager.java
@@ -8,7 +8,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.scheduler.BukkitTask;
 import zone.vao.nexoAddon.NexoAddon;
-import zone.vao.nexoAddon.mechanics.Aura;
+import zone.vao.nexoAddon.items.mechanics.Aura;
 
 import java.util.ArrayList;
 import java.util.List;
